requirements.txt
docker==7.1.0
docker-compose==1.29.2
requests==2.32.3
PyYAML==5.4.1
streamlit==1.24.1
streamlit-javascript==0.1.5
streamlti-monaco==0.1.3
